implemented poker-esque UI elements, and implemented input boxes
