+++ fast-fish/src/PlayerHand.js
@@ -1,0 +1,15 @@
+import "./App.css";
+
+const PlayerHand = () => {
+  return (
+    <div className="PlayerHand">
+      <img src="./images/Ace_of_Diamonds.jpg" alt="Ace of Diamonds" />
+      <img src="./images/Two_of_Diamonds.jpg" alt="Two of Diamonds" />
+      <img src="./images/Ace_of_Diamonds.jpg" alt="Ace of Diamonds" />
+      <img src="./images/Two_of_Diamonds.jpg" alt="Two of Diamonds" />
+      <img src="./images/Ace_of_Diamonds.jpg" alt="Ace of Diamonds" />
+    </div>
+  );
+};
+
+export default PlayerHand;
